+++ react/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Minor Changes
 
-- cae5ce6: Adds new programmatic actions to the React embed component for advanced integrations: goTo, createField, clearFields, getDocumentContent
+- cae5ce6: Adds new programmatic actions to the React embed component for advanced integrations: goTo, createField, removeFields, getDocumentContent